- Updates from review on pull request #2289

+++ addons/binding/org.openhab.binding.dsmr/src/main/java/org/openhab/binding/dsmr/internal/meter/DSMRMeterConstants.java
@@ -17,10 +17,16 @@ import java.util.concurrent.TimeUnit;
  * @since 2.1.0
  */
 public class DSMRMeterConstants {
-    // Timeout for receiving meter values (in seconds)
-    public static final int METER_VALUES_RECEIVED_TIMEOUT = (int) TimeUnit.SECONDS.toMillis(60);
-    // period between evaluations of meter values are received
-    public static final int METER_VALUES_TIMER_PERIOD = (int) TimeUnit.SECONDS.toMillis(10);
-    // unknown M-Bus channel
+    /**
+     * Timeout for receiving meter values (in seconds)
+     */
+    public static final long METER_VALUES_RECEIVED_TIMEOUT = TimeUnit.SECONDS.toMillis(60);
+    /**
+     * Period between evaluations of meter values are received
+     */
+    public static final long METER_VALUES_TIMER_PERIOD = TimeUnit.SECONDS.toMillis(10);
+    /**
+     * Unknown M-Bus channel
+     */
     public static final int UNKNOWN_CHANNEL = -1;
 }
